+++ src/main/java/ie/atu/GrpPoppies/CarPart/CustomerInterface.java
@@ -1,4 +1,4 @@
-package ie.atu.GrpPoppies.CarPart.Standard;
+package ie.atu.GrpPoppies.CarPart;
 
 public interface CustomerInterface {       //Basically adds abstraction meaning it doesn't care what it gets
     //String toString();                     //So basically this method turns everything to String (I think)
