+++ src/main/java/ie/atu/GrpPoppies/CarPart/DatabaseUtils.java
@@ -1,4 +1,4 @@
-package ie.atu.GrpPoppies.CarPart.Pool;
+package ie.atu.GrpPoppies.CarPart;
 
 import com.microsoft.sqlserver.jdbc.SQLServerDataSource;
