+++ src/main/java/ie/atu/GrpPoppies/CarPart/EnginePart.java
@@ -1,8 +1,6 @@
-package ie.atu.GrpPoppies.CarPart.Standard;
+package ie.atu.GrpPoppies.CarPart;
 
-import ie.atu.GrpPoppies.CarPart.Standard.CarPart;
-
-public class EnginePart extends CarPart {
+public class EnginePart extends CarPart{
     private String engineType;
     private double engineSize;
