+++ src/main/java/ie/atu/GrpPoppies/CarPart/EnginePartDB.java
@@ -1,7 +1,7 @@
-package ie.atu.GrpPoppies.CarPart.Pool;
+package ie.atu.GrpPoppies.CarPart;
 
-import ie.atu.GrpPoppies.CarPart.Standard.EnginePart;
 import java.sql.*;
+import java.util.ArrayList;
 import java.util.Scanner;
 
 import static java.lang.Integer.parseInt;
@@ -22,8 +22,27 @@ public class EnginePartDB {
         Epart.setEngineSize(engineSize);
         try {
             Connection conn = DatabaseUtils.getConnection();
-            String query = "INSERT INTO Engine_parts (part_number, name, manufacturer, supplier, quantity, " +
-                    "price, warranty, description, engine_type, engine_size) VALUES (?, ?, ?, ?, ?, ?, ?, ?,?,?)";
+
+            DatabaseMetaData metaData = conn.getMetaData();
+            ResultSet columns = metaData.getColumns(null, null, "Engine_parts", null);
+
+            ArrayList<String> columnNames = new ArrayList<>();
+
+            while (columns.next()) {
+                String columnName = columns.getString("COLUMN_NAME");
+                if (!columnName.equals("Internal_ID")) {// is the auto increment and will return
+                    columnNames.add(columnName);
+                }
+            }
+            String columnGet = String.join(",", columnNames);
+            String columnValues = "";
+            for (int i = 0; i < columnNames.size(); i++) {
+                columnValues += "?";//adds ? for every var
+                if (i != columnNames.size() - 1) {//stop extra , at end
+                    columnValues += ",";
+                }
+            }
+            String query = "INSERT INTO Engine_parts (" + columnGet + ") VALUES (" + columnValues + ")";
             PreparedStatement stmt = conn.prepareStatement(query);
 
             // set the parameters for the PreparedStatement
@@ -166,6 +185,7 @@ public class EnginePartDB {
             } else {
                 System.out.println("Car part not found.");
             }
+            conn.close();
         } catch (SQLException e) {
             System.out.println("Error Updating part: " + e.getMessage());
         }
