+++ src/main/java/ie/atu/GrpPoppies/CarPart/TyreInfoDB.java
@@ -1,9 +1,9 @@
-package ie.atu.GrpPoppies.CarPart.Pool;
-
-        import ie.atu.GrpPoppies.CarPart.Standard.TyreInfo;
+package ie.atu.GrpPoppies.CarPart;
 
         import java.sql.*;
+        import java.util.ArrayList;
         import java.util.Scanner;
+
         import static java.lang.Integer.parseInt;
 
 public class TyreInfoDB {
@@ -23,8 +23,27 @@ public class TyreInfoDB {
         try {
             // establish a connection to the database
             Connection conn = DatabaseUtils.getConnection();
-            String query = "INSERT INTO wheel_info (part_number, name, manufacturer, supplier, quantity, " +
-                    "price, warranty, description, tyretype, tyrerating, wheelsize) VALUES (?, ?, ?, ?, ?, ?, ?, ?,?,?,?)";;
+
+            DatabaseMetaData metaData = conn.getMetaData();
+            ResultSet columns = metaData.getColumns(null, null, "Wheel_info", null);
+
+            ArrayList<String> columnNames = new ArrayList<>();
+
+            while (columns.next()) {
+                String columnName = columns.getString("COLUMN_NAME");
+                if (!columnName.equals("Internal_ID")) {// is the auto increment and will return
+                    columnNames.add(columnName);
+                }
+            }
+            String columnGet = String.join(",", columnNames);
+            String columnValues = "";
+            for (int i = 0; i < columnNames.size(); i++) {
+                columnValues += "?";//adds ? for every var
+                if (i != columnNames.size() - 1) {//stop extra , at end
+                    columnValues += ",";
+                }
+            }
+            String query = "INSERT INTO wheel_info (" + columnGet + ") VALUES (" + columnValues + ")";
             PreparedStatement stmt = conn.prepareStatement(query);
 
             // set the parameters for the PreparedStatement
